pyproject.toml
[tool.black]
line-length = 88
skip-string-normalization = false
preview = false
skip-magic-trailing-comma = true
safe = true
# extend-exclude = '''
# /(
#   some_folder_here
#
# )/
# '''

[tool.isort]
py_version = "311"
profile = "black"
atomic = true # `true` cuando la versión de python target y que corre isort es la misma
lines_after_imports = 2
filter_files = true
skip_gitignore = true
combine_as_imports = true # Revisar
# known_third_party = []
known_first_party = ["users", "app", "back", "documents", "log", "questionnaires"] # Ajustar en cada proyecto

[build-system]
requires = ["setuptools==65.5.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sepraps"
version = "0.0.1"
authors = [
  { name="iCarto", email="info@icarto.es" },
]
readme = "README.md"
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Private :: Do Not Upload"
]
requires-python = "==3.11.*"
dependencies = [
    "Django==4.2",
    "django-cors-headers==3.14.0",
    "django-environ==0.10.0",
    "django-extensions==3.2.1",
    "django-filter==23.1",
    "django-spa==0.3.6",
    "djangorestframework==3.14.0",
    "djangorestframework-simplejwt==5.2.2",
    "djangorestframework-gis==1.0",
    "drf-extensions==0.7.1",
    "psycopg[binary]==3.1.8",
    "pandas==2.0.0",
    "django-debug-toolbar==4.0.0",
    "drf-spectacular==0.26.2",
    "sentry-sdk==1.20.0",
]



[tool.setuptools]
package-dir = {"sepraps" = "back"}
include-package-data = true

[tool.setuptools.package-data]
# *.css *.woff *.svg *.js *.gif *.ico *.png *.jpg *.html *.js.map
"sepraps.static" = ["**"]
"sepraps.templates" = ["**"]
